C 모듈의 calcsize/encode/decode 함수 반환형과 srclen 을 size_t로 통일

+++ src/hypua2jamo-c/hypua2jamo.h
@@ -5,33 +5,33 @@
 
 /* encoder */
 
-int hypua_c2d_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_c2d_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_c2d_ucs2_calcsize(const unsigned short *src, int srclen);
-ptrdiff_t hypua_c2d_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
-int hypua_p2jc_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_p2jc_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_p2jc_ucs2_calcsize(const unsigned short *src, int srclen);
-ptrdiff_t hypua_p2jc_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
-int hypua_p2jd_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_p2jd_ucs4_encode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_p2jd_ucs2_calcsize(const unsigned short *src, int srclen);
-ptrdiff_t hypua_p2jd_ucs2_encode(const unsigned short *src, int srclen, unsigned short *dst);
+size_t hypua_c2d_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_c2d_ucs4_encode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_c2d_ucs2_calcsize(const unsigned short *src, size_t srclen);
+size_t hypua_c2d_ucs2_encode(const unsigned short *src, size_t srclen, unsigned short *dst);
+size_t hypua_p2jc_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_p2jc_ucs4_encode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_p2jc_ucs2_calcsize(const unsigned short *src, size_t srclen);
+size_t hypua_p2jc_ucs2_encode(const unsigned short *src, size_t srclen, unsigned short *dst);
+size_t hypua_p2jd_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_p2jd_ucs4_encode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_p2jd_ucs2_calcsize(const unsigned short *src, size_t srclen);
+size_t hypua_p2jd_ucs2_encode(const unsigned short *src, size_t srclen, unsigned short *dst);
 
 /* decoder */
 
-int hypua_d2c_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_d2c_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_d2c_ucs2_calcsize(const unsigned short *src, int srclen);
-ptrdiff_t hypua_d2c_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
-int hypua_jc2p_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_jc2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_jc2p_ucs2_calcsize(const unsigned short *src, int srclen);
-ptrdiff_t hypua_jc2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
-int hypua_jd2p_ucs4_calcsize(const unsigned int *src, int srclen);
-ptrdiff_t hypua_jd2p_ucs4_decode(const unsigned int *src, int srclen, unsigned int *dst);
-int hypua_jd2p_calcsize_ucs2(const unsigned short *src, int srclen);
-ptrdiff_t hypua_jd2p_ucs2_decode(const unsigned short *src, int srclen, unsigned short *dst);
+size_t hypua_d2c_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_d2c_ucs4_decode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_d2c_ucs2_calcsize(const unsigned short *src, size_t srclen);
+size_t hypua_d2c_ucs2_decode(const unsigned short *src, size_t srclen, unsigned short *dst);
+size_t hypua_jc2p_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_jc2p_ucs4_decode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_jc2p_ucs2_calcsize(const unsigned short *src, size_t srclen);
+size_t hypua_jc2p_ucs2_decode(const unsigned short *src, size_t srclen, unsigned short *dst);
+size_t hypua_jd2p_ucs4_calcsize(const unsigned int *src, size_t srclen);
+size_t hypua_jd2p_ucs4_decode(const unsigned int *src, size_t srclen, unsigned int *dst);
+size_t hypua_jd2p_calcsize_ucs2(const unsigned short *src, size_t srclen);
+size_t hypua_jd2p_ucs2_decode(const unsigned short *src, size_t srclen, unsigned short *dst);
 
 /* decoder generic */
 
@@ -47,22 +47,22 @@ void hypua_decoder_init(
 );
 int hypua_decoder_getstate(void *decoder);
 int hypua_decoder_setstate(void *decoder, int state);
-int hypua_decoder_calcsize_ucs2(void *decoder, void *src, int srclen);
-int hypua_decoder_calcsize_ucs4(void *decoder, void *src, int srclen);
-int hypua_decoder_calcsize_flush(void *t);
-ptrdiff_t hypua_decoder_decode_ucs2(
+size_t hypua_decoder_calcsize_ucs2(void *decoder, void *src, size_t srclen);
+size_t hypua_decoder_calcsize_ucs4(void *decoder, void *src, size_t srclen);
+size_t hypua_decoder_calcsize_flush(void *t);
+size_t hypua_decoder_decode_ucs2(
 		void *decoder,
 		void *src,
-		int srclen,
+		size_t srclen,
 		void *dst
 );
-ptrdiff_t hypua_decoder_decode_ucs4(
+size_t hypua_decoder_decode_ucs4(
 		void *decoder,
 		void *src,
-		int srclen,
+		size_t srclen,
 		void *dst
 );
-ptrdiff_t hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
-ptrdiff_t hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
+size_t hypua_decoder_decode_flush_ucs2(void *decoder, void *dst);
+size_t hypua_decoder_decode_flush_ucs4(void *decoder, void *dst);
 
 #endif
